handwriting_leo/requirement.txt
matplotlib==3.4.2
Pillow==8.0.1
torch==1.7.0
torchvision==0.8.1
torchaudio==0.7.0
scikit-learn==0.24.2
pandas==0.25.3